Refactor: 멤버 관리 table 추상화
