+++ frontend/src/components/common/index.ts
@@ -9,3 +9,4 @@ export * from './spinner'
 export * from './image-input'
 export * from './icon-button'
 export * from './intersection-observer-loader'
+export * from './table'
