+++ frontend/src/components/common/table/content/index.ts
@@ -1,0 +1,1 @@
+export { TableContent } from './Content'
