+++ frontend/src/components/common/table/filter/index.ts
@@ -1,0 +1,2 @@
+export { TableFilter } from './Filter'
+export { TableFilterSkeleton } from './Filter'
