+++ frontend/src/components/common/table/index.ts
@@ -1,0 +1,3 @@
+export * from './content'
+export * from './filter'
+export * from './pagination'
