+++ frontend/src/components/common/table/pagination/index.ts
@@ -1,0 +1,1 @@
+export { TablePagination } from './Pagination'
