+++ frontend/src/components/feature/admin/index.ts
@@ -1,0 +1,1 @@
+export * from './member-table'
