+++ frontend/src/components/feature/admin/member-table/index.ts
@@ -1,0 +1,3 @@
+export * from './not-found-table'
+export * from './table'
+export * from './table-skeleton'
