+++ frontend/src/components/feature/admin/member-table/not-found-table/index.ts
@@ -1,0 +1,1 @@
+export { NotFoundMemberTable } from './NotFoundTable'
