+++ frontend/src/components/feature/admin/member-table/table-skeleton/TableSkeleton.tsx
@@ -6,10 +6,8 @@ import {
   useReactTable,
 } from '@tanstack/react-table'
 
+import { TableContent, TableFilterSkeleton } from '@/components/common'
 import { Skeleton } from '@/components/ui'
-
-import { MemberTableContent } from './content'
-import { MemberTableFilterSkeleton } from './filter/Filter'
 
 interface MemberTableSkeletonProps<T> {
   columns: ColumnDef<T>[]
@@ -28,8 +26,8 @@ export function MemberTableSkeleton<T>({
 
   return (
     <div className="flex w-full max-w-screen-lg flex-col gap-2 pb-20">
-      <MemberTableFilterSkeleton />
-      <MemberTableContent table={table} />
+      <TableFilterSkeleton />
+      <TableContent table={table} />
       <div className="flex flex-col gap-2">
         <Skeleton className="my-2 h-6 w-full" />
         <Skeleton className="my-2 h-6 w-full" />
