+++ frontend/src/components/feature/admin/member-table/table-skeleton/index.ts
@@ -1,0 +1,1 @@
+export { MemberTableSkeleton } from './TableSkeleton'
