+++ frontend/src/components/feature/admin/member-table/table/index.ts
@@ -1,0 +1,1 @@
+export { MemberTable } from './MemberTable'
