+++ frontend/src/components/feature/index.ts
@@ -2,3 +2,4 @@ export * from './layout'
 export * from './post'
 export * from './error-fallback'
 export * from './comment'
+export * from './admin'
