+++ frontend/src/pages/admin/member/components/approve-member-table/index.ts
@@ -1,0 +1,1 @@
+export { ApproveMemberTable } from './ApproveMemberTable'
