+++ frontend/src/pages/admin/member/components/change-role-table/index.ts
@@ -1,0 +1,1 @@
+export { ChangeRoleTable } from './ChangRoleTable'
