+++ frontend/src/pages/admin/member/components/expel-member-table/ExpelMemberTable.tsx
@@ -2,24 +2,25 @@ import { useQuery } from '@tanstack/react-query'
 import { ColumnDef } from '@tanstack/react-table'
 
 import { NotFound } from '@/components/common'
-import { AdminUserQuries } from '@/service/api'
-import { UserResponseDto } from '@/service/model'
-
-import { ChangeRoleDialog } from './change-role-dialog'
 import {
   MemberTable,
-  MemberTableNone,
   MemberTableSkeleton,
-} from './member-table'
+  NotFoundMemberTable,
+} from '@/components/feature'
+import { AdminUserQuries } from '@/service/api'
+import { UserResponseDto } from '@/service/model'
+import { convertRoleName } from '@/utils'
 
-export const ChangeRoleTable = () => {
+import { ExpelMemberDialog } from './expel-member-dialog'
+
+export const ExpelMemberTable = () => {
   const {
     data: activeUsers,
     status,
     error,
   } = useQuery(AdminUserQuries.active())
 
-  const changeRoleColumn: ColumnDef<UserResponseDto>[] = [
+  const expelMemberColumn: ColumnDef<UserResponseDto>[] = [
     {
       header: '',
       id: 'id',
@@ -46,12 +47,19 @@ export const ChangeRoleTable = () => {
     {
       accessorKey: 'role',
       header: '등급',
+      cell: ({ row }) => (
+        <div className="text-center">
+          {convertRoleName(row.getValue('role'))}
+        </div>
+      ),
+    },
+    {
+      accessorKey: 'isBanned',
+      header: '',
       cell: ({ row }) => {
-        const user = row.original
-
         return (
           <div className="flex justify-center">
-            <ChangeRoleDialog user={user} />
+            <ExpelMemberDialog user={row.original} />
           </div>
         )
       },
@@ -59,15 +67,18 @@ export const ChangeRoleTable = () => {
   ]
 
   if (status === 'pending')
-    return <MemberTableSkeleton columns={changeRoleColumn} />
+    return <MemberTableSkeleton columns={expelMemberColumn} />
 
   if (error) return <NotFound />
 
   if (!activeUsers.length) {
     return (
-      <MemberTableNone message="회원이 없습니다." columns={changeRoleColumn} />
+      <NotFoundMemberTable
+        message="회원이 없습니다."
+        columns={expelMemberColumn}
+      />
     )
   }
 
-  return <MemberTable data={activeUsers} columns={changeRoleColumn} />
+  return <MemberTable data={activeUsers} columns={expelMemberColumn} />
 }
