+++ frontend/src/pages/admin/member/components/expel-member-table/index.ts
@@ -1,0 +1,1 @@
+export { ExpelMemberTable } from './ExpelMemberTable'
