--- frontend/src/pages/admin/member/components/member-table/content/index.ts
@@ -1,1 +1,0 @@
-export { MemberTableContent } from './Content'
