--- frontend/src/pages/admin/member/components/member-table/filter/index.ts
@@ -1,1 +1,0 @@
-export { MemberTableFilter } from './Filter'
