--- frontend/src/pages/admin/member/components/member-table/index.ts
@@ -1,3 +1,0 @@
-export { MemberTable } from './MemberTable'
-export { MemberTableNone } from './MemberTableNone'
-export { MemberTableSkeleton } from './MemberTableSkeleton'
