--- frontend/src/pages/admin/member/components/member-table/pagination/index.ts
@@ -1,1 +1,0 @@
-export { MemberTablePagination } from './Pagination'
